deploy/requirements-server.txt
alembic==1.16.5
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
email-validator==2.2.0
Flask==3.1.2
Flask-Cors==6.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
Pillow==11.3.0
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.0.1
python-pptx==1.0.2
requests==2.32.5
SQLAlchemy==2.0.43
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
XlsxWriter==3.2.8
